make plain tree output the root command
